ai_agents_cookbooks/human-in-the-loop/pyproject.toml
[project]
name = "human-in-loop"
version = "0.1.0"
description = ""
authors = [
    {name = "Homayoun"}
]
readme = "README.md"
requires-python = "<4.0,>=3.12"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langgraph (>=0.5.3,<0.6.0)",
    "langmem (>=0.0.28,<0.0.29)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-core (>=0.3.69,<0.4.0)",
    "tavily-python (>=0.7.10,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
